move para_scheduler back to contrib
