+++ ignite/contrib/handlers/__init__.py
@@ -10,3 +10,12 @@ from ignite.contrib.handlers.trains_logger import TrainsLogger
 from ignite.contrib.handlers.visdom_logger import VisdomLogger
 from ignite.contrib.handlers.wandb_logger import WandBLogger
 from ignite.handlers import global_step_from_engine  # ref
+from ignite.handlers.param_scheduler import (
+    ConcatScheduler,
+    CosineAnnealingScheduler,
+    LinearCyclicalScheduler,
+    LRScheduler,
+    ParamGroupScheduler,
+    PiecewiseLinear,
+    create_lr_scheduler_with_warmup,
+)
