+++ ignite/contrib/handlers/param_scheduler.py
