+++ tests/ignite/contrib/handlers/test_param_scheduler.py
@@ -5,8 +5,7 @@ import pytest
 import torch
 from torch.optim.lr_scheduler import ExponentialLR, StepLR
 
-from ignite.engine import Engine, Events
-from ignite.handlers.param_scheduler import (
+from ignite.contrib.handlers.param_scheduler import (
     ConcatScheduler,
     CosineAnnealingScheduler,
     LinearCyclicalScheduler,
@@ -16,6 +15,7 @@ from ignite.handlers.param_scheduler import (
     PiecewiseLinear,
     create_lr_scheduler_with_warmup,
 )
+from ignite.engine import Engine, Events
 from tests.ignite.contrib.handlers import MockFP16DeepSpeedZeroOptimizer
 
 try:
